+++ docs/ta2-notes/test_search.md
@@ -1,0 +1,7 @@
+
+
+```
+{"userAgent":"TwoRavens","version":"2018.7.7","timeBound":4,"priority":1,"allowedValueTypes":["DATASET_URI","CSV_URI"],"problem":{"problem":{"id":"185_bl_problem_TRAIN","version":"1.0","name":"NULL","description":"","taskType":"CLASSIFICATION","taskSubtype":"MULTICLASS","performanceMetrics":[{"metric":"F1_MACRO"}]},"inputs":[{"datasetId":"185_bl_dataset_TRAIN","targets":[{"resourceId":"0","columnIndex":17,"columnName":"Hall_of_Fame"}]}]},"template":{"inputs":[],"outputs":[],"steps":[]},"inputs":[{"dataset_uri":"file:///ravens_volume/test_data/185_baseball/TRAIN/dataset_TRAIN/datasetDoc.json"}]}
+```
+
+{"searchId":"9ecf53db-af0e-4551-a97f-a4c59d92c5b0"}
