+++ requirements/base.txt
@@ -9,6 +9,11 @@ django-webpack-loader==0.5.0
 jsonfield==2.0.2
 requests==2.18.4
 
+# django channels
+#
+channels==2.1.5
+
+
 # fabric for dev: to run npm and test server with one command
 #
 Fabric3==1.13.1.post1
