+++ tworaven_apps/configurations/urls.py
@@ -50,11 +50,5 @@ urlpatterns = (
         views.view_get_problem_data_info,
         name='view_get_problem_data_info'),
 
-    #url(r'^(?P<app_name_in_url>\w{5,25})$',
-    #    views.view_rook_route,
-    #    name='view_rook_route'),
 
-    #url(r'^web-pack$',
-    #    views_test.view_webpack_test,
-    #    name='view_webpack_test'),
 )
