+++ tworaven_apps/websocket_views/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class WsTestConfig(AppConfig):
-    name = 'ws_test'
+    name = 'websocket_views'
