+++ tworaven_apps/websocket_views/routing.py
@@ -3,6 +3,6 @@ from django.conf.urls import url
 from . import consumers
 
 websocket_urlpatterns = [
-    url(r'^ws/chat/(?P<room_name>[^/]+)/$', consumers.ChatConsumer),
+    url(r'^ws/connect/(?P<room_name>[^/]+)/$', consumers.ChatConsumer),
 
 ]
