+++ tworaven_apps/websocket_views/templates/ws_test/xindex.html
@@ -20,7 +20,7 @@
 
         document.querySelector('#room-name-submit').onclick = function(e) {
             var roomName = document.querySelector('#room-name-input').value;
-            window.location.pathname = 'ws-test/' + roomName + '/';
+            window.location.pathname = 'ws-views/' + roomName + '/';
         };
     </script>
 </body>
