+++ tworaven_apps/websocket_views/urls.py
@@ -1,0 +1,13 @@
+# chat/urls.py
+from django.conf.urls import url
+
+from . import views
+
+urlpatterns = [
+
+    # Test urls...
+    #url(r'alarm-ok', views.view_alarm, name='view_alarm'),
+    #url(r'^(?P<room_name>[^/]+)/$', views.room, name='room'),
+    #url(r'^$', views.index, name='index'),
+
+]
