+++ tworavensproject/routing.py
@@ -1,12 +1,12 @@
 from channels.auth import AuthMiddlewareStack
 from channels.routing import ProtocolTypeRouter, URLRouter
-import tworaven_apps.ws_test.routing
+import tworaven_apps.websocket_views.routing
 
 application = ProtocolTypeRouter({
     # (http->django views is added by default)
     'websocket': AuthMiddlewareStack(
         URLRouter(
-            tworaven_apps.ws_test.routing.websocket_urlpatterns
+            tworaven_apps.websocket_views.routing.websocket_urlpatterns
         )
     ),
 })
